Composition refactor in FileGridItem and FileListItem (DriveItem views)
