+++ src/components/FileExplorer/FileExplorerItem/FileGridItem/FileGridItem.scss
