+++ src/components/FileExplorer/FileExplorerItem/FileListItem/FileListItem.scss
