--- src/components/FileExplorer/FilesGrid/FileGridItem/FileGridItem.tsx
@@ -1,333 +1,0 @@
-import React, { MouseEvent, Fragment, ReactNode } from 'react';
-import Dropdown from 'react-bootstrap/Dropdown';
-import * as Unicons from '@iconscout/react-unicons';
-
-import FileDropdownActions from '../../../dropdowns/FileDropdownActions/FileDropdownActions';
-import { storageActions, storageSelectors, storageThunks } from '../../../../store/slices/storage';
-
-import folderService from '../../../../services/folder.service';
-import fileService from '../../../../services/file.service';
-import { AppDispatch, RootState } from '../../../../store';
-import { connect } from 'react-redux';
-import { DriveFileMetadataPayload, DriveFolderMetadataPayload, DriveItemData, FolderPath, UserSettings } from '../../../../models/interfaces';
-
-import './FileGridItem.scss';
-import { ItemAction, Workspace } from '../../../../models/enums';
-
-import './FileGridItem.scss';
-import iconService from '../../../../services/icon.service';
-import { uiActions } from '../../../../store/slices/ui';
-import { getItemFullName } from '../../../../services/storage.service/storage-name.service';
-import { getAllItems } from '../../../../services/dragAndDrop.service';
-import { NativeTypes } from 'react-dnd-html5-backend';
-import { ConnectDropTarget, DropTarget, DropTargetCollector, DropTargetSpec } from 'react-dnd';
-interface FileGridItemProps {
-  user: UserSettings;
-  item: DriveItemData;
-  selectedItems: DriveItemData[];
-  currentFolderId: number;
-  namePath: FolderPath[];
-  isItemSelected: (item: DriveItemData) => boolean;
-  isDriveItemInfoMenuOpen: boolean;
-  dispatch: AppDispatch;
-  workspace: Workspace;
-  isOver: boolean;
-  connectDropTarget: ConnectDropTarget;
-}
-
-interface FileGridItemState {
-  isEditingName: boolean;
-  dirtyName: string;
-  itemRef: React.RefObject<HTMLDivElement>;
-  nameInputRef: React.RefObject<HTMLInputElement>;
-  oldValueIsDriveItemInfoMenuOpen: boolean;
-}
-
-class FileGridItem extends React.Component<FileGridItemProps, FileGridItemState> {
-  constructor(props: FileGridItemProps) {
-    super(props);
-
-    this.state = {
-      isEditingName: false,
-      dirtyName: '',
-      itemRef: React.createRef(),
-      nameInputRef: React.createRef(),
-      oldValueIsDriveItemInfoMenuOpen: props.isDriveItemInfoMenuOpen
-    };
-  }
-
-  componentDidMount() {
-    this.updateHeight();
-
-    window.addEventListener('resize', this.updateHeight);
-  }
-
-  componentWillUnmount() {
-    window.removeEventListener('resize', this.updateHeight);
-  }
-
-  updateHeight = () => {
-    this.forceUpdate();
-  }
-
-  get nameNode(): JSX.Element {
-    const { item } = this.props;
-    const { isEditingName, dirtyName, nameInputRef } = this.state;
-    const ṣpanDisplayClass: string = !isEditingName ? 'block' : 'hidden';
-
-    return (
-      <Fragment>
-        <div className={isEditingName ? 'flex' : 'hidden'}>
-          <input
-            className="w-full dense border border-white no-ring rect"
-            onClick={(e) => e.stopPropagation()}
-            ref={nameInputRef}
-            type="text"
-            value={dirtyName}
-            placeholder="Name"
-            onChange={this.onNameChanged}
-            onBlur={this.onNameBlurred}
-            onKeyPress={this.onEnterKeyPressed}
-            autoFocus
-          />
-          <span className="ml-1">{item.type ? ('.' + item.type) : ''}</span>
-        </div>
-        <span
-          className={`${ṣpanDisplayClass} file-grid-item-name-span`}
-          onClick={(e) => e.stopPropagation()}
-          onDoubleClick={this.onNameDoubleClicked}
-        >{getItemFullName(item.name, item.type)}</span>
-      </Fragment>
-    );
-  }
-
-  confirmNameChange() {
-    const { item } = this.props;
-    const { dirtyName, nameInputRef } = this.state;
-    const data: DriveFileMetadataPayload | DriveFolderMetadataPayload = { metadata: { itemName: dirtyName } };
-    const isTeam = this.props.workspace === Workspace.Business ? true : false;
-
-    try {
-      const oldName = item.name;
-
-      if (item.name !== dirtyName) {
-        this.props.dispatch(storageActions.resetItemName({ name: dirtyName, id: item.id, isFolder: !!item.isFolder }));
-        if (item.isFolder) {
-          folderService.updateMetaData(item.id, data, isTeam)
-            .then((res) => {
-              if (!res) {
-                this.props.dispatch(storageActions.resetItemName({ name: oldName, id: item.id, isFolder: !!item.isFolder }));
-              }
-              /*
-              this.props.dispatch(
-                storageThunks.fetchFolderContentThunk()
-              );
-              */
-            });
-        } else {
-          fileService.updateMetaData(item.fileId, data as DriveFileMetadataPayload, isTeam).then((res) => {
-            if (!res) {
-              this.props.dispatch(storageActions.resetItemName({ name: oldName, id: item.id, isFolder: !!item.isFolder }));
-            }
-            /*
-            this.props.dispatch(
-              storageThunks.fetchFolderContentThunk()
-            );
-            */
-          });
-        }
-      }
-    } catch (e) {
-      console.log('Error during folder/file name change: ', e);
-    } finally {
-      nameInputRef.current?.blur();
-    }
-  }
-
-  onNameDoubleClicked = (e: MouseEvent): void => {
-    const { item } = this.props;
-    const { nameInputRef } = this.state;
-
-    e.stopPropagation();
-
-    this.setState(
-      { isEditingName: true, dirtyName: item.name },
-      () => nameInputRef.current?.focus()
-    );
-  }
-
-  onNameBlurred = (): void => {
-    this.setState({ isEditingName: false });
-  }
-
-  onNameChanged = (e: React.ChangeEvent<HTMLInputElement>): void => {
-    this.setState({ dirtyName: e.target.value });
-  }
-
-  onEnterKeyPressed = (e: React.KeyboardEvent): void => {
-    if (e.key === 'Enter') {
-      this.confirmNameChange();
-    }
-  }
-
-  onRenameButtonClicked = (): void => {
-    const { item } = this.props;
-    const { nameInputRef } = this.state;
-
-    this.setState(
-      { isEditingName: true, dirtyName: item.name },
-      () => setTimeout(() => nameInputRef.current && nameInputRef.current.focus(), 0)
-    );
-  }
-
-  onDownloadButtonClicked = (e: MouseEvent): void => {
-    const { item, dispatch } = this.props;
-
-    e.stopPropagation();
-
-    dispatch(storageThunks.downloadItemsThunk([item]));
-  }
-
-  onShareButtonClicked = (): void => {
-    const { dispatch, item } = this.props;
-
-    dispatch(storageActions.setItemToShare(item.id));
-    dispatch(uiActions.setIsShareItemDialogOpen(true));
-  }
-
-  onInfoButtonClicked = (): void => {
-    this.props.dispatch(storageActions.setInfoItem(this.props.item.id));
-    this.props.dispatch(uiActions.setIsDriveItemInfoMenuOpen(true));
-  }
-
-  onDeleteButtonClicked = (): void => {
-    const { dispatch, item } = this.props;
-
-    dispatch(storageActions.setItemsToDelete([item]));
-    dispatch(uiActions.setIsDeleteItemsDialogOpen(true));
-  }
-
-  onItemClicked = (): void => {
-    const { item, dispatch, isItemSelected } = this.props;
-
-    if (!item.isFolder) {
-      isItemSelected(item) ?
-        dispatch(storageActions.deselectItems([item])) :
-        dispatch(storageActions.selectItems([item]));
-    }
-  }
-
-  onItemRightClicked = (e: MouseEvent): void => {
-    e.preventDefault();
-  }
-
-  onItemDoubleClicked = (): void => {
-    const { dispatch, item } = this.props;
-
-    if (item.isFolder) {
-      dispatch(storageThunks.goToFolderThunk({ name: item.name, id: item.id }));
-    }
-  }
-
-  render(): ReactNode {
-    const { itemRef } = this.state;
-    const { item, isItemSelected, isOver, connectDropTarget } = this.props;
-    const selectedClassNames: string = isItemSelected(item) ? 'selected' : '';
-    const ItemIconComponent = iconService.getItemIcon(item.isFolder, item.type);
-    const height = this.state.itemRef.current ?
-      this.state.itemRef.current?.clientWidth + 'px' :
-      'auto';
-
-    return connectDropTarget(
-      <div
-        ref={itemRef}
-        style={{ height }}
-        className={`${selectedClassNames} ${isOver ? 'drag-over-effect' : ''} group file-grid-item`}
-        onContextMenu={this.onItemRightClicked}
-        onClick={this.onItemClicked}
-        onDoubleClick={this.onItemDoubleClicked}
-        draggable={false}
-      >
-        <Dropdown>
-          <Dropdown.Toggle variant="success" id="dropdown-basic" className="file-grid-item-actions-button">
-            <Unicons.UilEllipsisH className="w-full h-full" />
-          </Dropdown.Toggle>
-          <Dropdown.Menu>
-            <FileDropdownActions
-              hiddenActions={item.isFolder ? [ItemAction.Download, ItemAction.Share] : []}
-              onRenameButtonClicked={this.onRenameButtonClicked}
-              onDownloadButtonClicked={this.onDownloadButtonClicked}
-              onShareButtonClicked={this.onShareButtonClicked}
-              onInfoButtonClicked={this.onInfoButtonClicked}
-              onDeleteButtonClicked={this.onDeleteButtonClicked}
-            />
-          </Dropdown.Menu>
-        </Dropdown>
-        <div className="file-grid-item-icon-container">
-          <ItemIconComponent className="file-icon m-auto" />
-        </div>
-        <div className="text-center mt-3">
-          <div className="mb-1">
-            {this.nameNode}
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
-
-const dropTargetSpec: DropTargetSpec<FileGridItemProps> = {
-  drop: (props, monitor, component) => {
-    const { dispatch, namePath, item } = props;
-    const droppedData = monitor.getItem();
-
-    if (item.isFolder) {
-      const namePathDestinationArray = namePath.map(level => level.name);
-
-      namePathDestinationArray[0] = '';
-
-      let folderPath = namePathDestinationArray.join('/');
-
-      folderPath = folderPath + '/' + item.name;
-
-      getAllItems(droppedData, folderPath).then(async ({ rootList, files }) => {
-        if (files) {
-          // files where dragged directly
-          await dispatch(storageThunks.uploadItemsThunk({ files, parentFolderId: item.id, folderPath }));
-        }
-        if (rootList) {
-          for (const root of rootList) {
-            const currentFolderId = item.id;
-
-            await dispatch(storageThunks.createFolderTreeStructureThunk({ root, currentFolderId }));
-          }
-        }
-      });
-    }
-  },
-  hover: (props, monitor, component) => { }
-};
-
-const dropTargetCollect: DropTargetCollector<{ isOver: boolean, connectDropTarget: ConnectDropTarget }, FileGridItemProps> = (connect, monitor, props) => {
-  const isOver = monitor.isOver() && props.item.isFolder;
-
-  return {
-    isOver,
-    connectDropTarget: connect.dropTarget()
-  };
-};
-
-export default connect(
-  (state: RootState) => {
-    const isItemSelected = storageSelectors.isItemSelected(state);
-    const currentFolderId = storageSelectors.currentFolderId(state);
-
-    return {
-      isDriveItemInfoMenuOpen: state.ui.isDriveItemInfoMenuOpen,
-      selectedItems: state.storage.selectedItems,
-      namePath: state.storage.namePath,
-      currentFolderId,
-      isItemSelected,
-      workspace: state.team.workspace
-    };
-  })(DropTarget((props) => props.item.isFolder ? [NativeTypes.FILE] : [], dropTargetSpec, dropTargetCollect)(FileGridItem));
